+++ JUN_All/tools/A00080_KWI_creator_V03/app/core/0010_src/A0202_Src_LDA_constraint_entry.py
@@ -1,0 +1,1 @@
+(BoneReference1=(BoneName="{{BONE_1}}"),BoneReference2=(BoneName="{{BONE_2}}"))